[thymeleaf] Make fragments usable without view folder prefix
This commit enhances the DefaultTemplateEngineProducer in a way that it
is possible to use Thymeleaf Fragments as they are shown in the
documentation.

Before this commit, only views returned by controllers were directly
found, because they had the '/WEB-INF/views/' prefix already added to
their path. Fragments which are called inside those views, like by using
'<div th:replace="fragments :: header"></div>', lead to an error, because
they were tried to resolve on the context root. One solution was to
clutter up those declarations by adding the view directory manually, but
this was error prone since the view directory might be changed by
configuration.

Now a second template resolver is checking for other files which are not
returned by controllers BUT are existent inside the view directory. This
way, the fragments can be used without using prefixes or suffixes.